requirements.txt
Flask==3.0.3
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
